refactor: functionsの関数を役割ごとに切り出し、見通しをよくした
