+++ apps/functions/src/types/DownloadAudioResult.ts
@@ -1,0 +1,3 @@
+export type DownloadAudioResult = {
+  tempFilePath: string;
+};
